split  GiphySnapshotTests
try to fix random fail snapshot test by wrapping GiphyConfirmationViewController in a simple navigation controller
