+++ Wire-iOS Tests/ReferenceImages_64/Wire_iOS_Tests.GiphySearchViewControllerSnapshotTests/testEmptySearchScreenWithKeyword_{375, 667}@2x.png
